backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
celery[redis]
python-dotenv
pydantic
requests
